Refactor object deserialization to YAML and fix various bugs
- Replace manual ParseValues/DeserializeValues with YamlDotNet deserialization and [YamlMember] aliases across all block types
- Add Vector3ConverterYaml for YAML vector serialization
- Rewrite TeleporterHandler on top of TriggerHandler with pickup/projectile support and minimum forced cooldowns
- Fix locker chamber population and add LockerType.None
- Add InteractionObject cache and clean up event handlers on destroy
- Make LogManager thread-safe and bounded
- Spawn schematics on a background task to prevent client crashes on large schematics
- Harden Updater against zip path traversal and fix UnityWebRequest disposal
- Fix block executor variable scoping, recursive executor assignment and for-loop bounds
- Add null guards for colliders/LOD zones and clean up schematic transforms on destroy
- Batch sync scheduling and snapshot processing in SyncManager

+++ ThaumielMapEditor/API/Helpers/BlockParser/MECBlocks.cs
@@ -59,7 +59,7 @@ namespace ThaumielMapEditor.API.Helpers.BlockParser
 
     public class WaitForSeconds : BlockBase
     {
-        public uint WaitTime { get; set; }
+        public float WaitTime { get; set; }
 
         public List<object?> Stack { get; set; } = [];
